Finish collapse
