--- packages/web/src/Collapse/index.js
@@ -1,1 +1,0 @@
-export { default } from './Collapse';
