--- packages/web/src/List/index.js
@@ -1,3 +1,0 @@
-export { default as ListGroup } from './ListGroup';
-export { default as ListItem } from './ListItem';
-export { ListContent } from './ListContent';
