+++ packages/web/src/ListGroup/index.js
@@ -1,1 +1,3 @@
-export { default } from './ListGroup';
+export { default as ListGroup } from './ListGroup';
+export { default as ListItem } from './ListItem';
+export { ListContent } from './ListContent';
