--- packages/web/src/ListGroup/type.js
@@ -1,15 +1,0 @@
-// @flow
-
-import type { Node } from 'react';
-
-export type ListItemType = {
-  content: Node,
-  badge?: string | number,
-  href?: string,
-  id?: string,
-  children?: Node | Array<ListItemType>,
-  open?: boolean,
-  childrenHeight?: number
-};
-
-export type ThemeTyped = 'light' | 'dark';
